Refactor editor hot paths
