+++ negative2positive/index.html
@@ -2622,11 +2622,6 @@
     }
   </style>
 
-  <!-- External Libraries -->
-  <script type="module">
-    import LibRaw from 'libraw-wasm';
-    window.LibRaw = LibRaw;
-  </script>
 </head>
 
 <body>
