+++ negative2positive/src/silvercore/engine/filmPresetsLoader.js
@@ -1,0 +1,13 @@
+let filmPresetsPromise = null;
+
+export function loadFilmPresets() {
+  if (!filmPresetsPromise) {
+    filmPresetsPromise = import('./FilmPresets.js')
+      .then(({ filmPresets }) => filmPresets || {})
+      .catch((err) => {
+        filmPresetsPromise = null;
+        throw err;
+      });
+  }
+  return filmPresetsPromise;
+}
